pyproject.toml
[project]
name = "openreview-mcp"
version = "0.1.0"
description = "MCP server for OpenReview — search submissions, fetch reviews, meta-reviews, rebuttals, and decisions from NeurIPS, ICLR, ACL ARR, COLM, TMLR, and more."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "OpenCódice Research", email = "hello@opencodice.org" },
]
keywords = ["mcp", "openreview", "academic", "research", "peer-review", "neurips", "iclr", "acl"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "openreview-py>=1.38.0",
    "pydantic>=2.6.0",
    "diskcache>=5.6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "vcrpy>=6.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[project.scripts]
openreview-mcp = "openreview_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/OpenCodice-Research/openreview-mcp"
Repository = "https://github.com/OpenCodice-Research/openreview-mcp"
Issues = "https://github.com/OpenCodice-Research/openreview-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openreview_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
